fix: unify styles to one hook
